+++ packages/legacy/urql/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @graphql-mesh/urql-exchange
 
+## 0.105.11
+
+### Patch Changes
+
+- [#8206](https://github.com/ardatan/graphql-mesh/pull/8206)
+  [`36dc08b`](https://github.com/ardatan/graphql-mesh/commit/36dc08b0fd06a65301d3b10b3403cd90e40f5ee1)
+  Thanks [@renovate](https://github.com/apps/renovate)! - dependencies updates:
+
+  - Added dependency
+    [`value-or-promise@^1.0.12` ↗︎](https://www.npmjs.com/package/value-or-promise/v/1.0.12) (to
+    `dependencies`)
+
+- [#8206](https://github.com/ardatan/graphql-mesh/pull/8206)
+  [`36dc08b`](https://github.com/ardatan/graphql-mesh/commit/36dc08b0fd06a65301d3b10b3403cd90e40f5ee1)
+  Thanks [@renovate](https://github.com/apps/renovate)! - Improvements on Executor typings
+
+- Updated dependencies
+  [[`36dc08b`](https://github.com/ardatan/graphql-mesh/commit/36dc08b0fd06a65301d3b10b3403cd90e40f5ee1)]:
+  - @graphql-mesh/runtime@0.105.11
+
 ## 0.105.10
 
 ### Patch Changes